+++ src/components/Constant.js
@@ -1,0 +1,37 @@
+export const CategoryList = [
+  {
+    CategoryName: "French",
+    id: 1,
+  },
+  {
+    CategoryName: "Italian",
+    id: 2,
+  },
+  {
+    CategoryName: "Irish",
+    id: 3,
+  },
+  {
+    CategoryName: "Animal",
+    id: 4,
+  },
+  {
+    CategoryName: "Flower",
+    id: 5,
+  },
+  {
+    CategoryName: "Christmas",
+    id: 6,
+  },
+  {
+    CategoryName: "Valentine's",
+    id: 7,
+  },
+];
+
+export const steps = [
+  "Customer Details",
+  "Shipping Details",
+  "Payment Details",
+  "Confirm Order",
+];
