+++ src/components/ShippingDetailsCases/Shipping Details.js
@@ -1,0 +1,180 @@
+import { useState } from "react";
+import { Box, Stepper, Step, Button, StepLabel } from "@mui/material";
+import { Link } from "react-router-dom";
+import { Case0 } from "./Case0";
+import { Case1 } from "./Case1";
+import { Case2 } from "./Case2";
+import { Case3 } from "./Case3";
+import { steps } from "../Constant";
+import "./ShippingCases.css";
+
+
+export function ShippingDetails() {
+  const [activeStep, setActiveStep] = useState(0);
+
+  const [validationError, setValidationError] = useState("");
+
+  const [customerInformation, setCustomerInformation] = useState([]);
+
+  const TotalPrice = JSON.parse(localStorage.getItem("totalPrice"));
+
+  const [discountedPrice, setDiscountedPrice] = useState(TotalPrice);
+
+  const handleVoucher = () => {
+    setDiscountedPrice(TotalPrice - 30 / 100);
+  };
+
+  const CartItems = JSON.parse(localStorage.getItem("cartItems"));
+
+  const nextStep = () => {
+    if (activeStep === 0) {
+      if (
+        customerInformation.fullName === "" ||
+        customerInformation.email === "" ||
+        customerInformation.phoneNumber === ""
+      ) {
+        setValidationError("Please fill all the fields");
+      } else {
+        setValidationError("");
+        setActiveStep(activeStep + 1);
+      }
+    }
+    if (activeStep === 1) {
+      if (
+        customerInformation.city === "" ||
+        customerInformation.zipCode === "" ||
+        customerInformation.country === "" ||
+        customerInformation.province === ""
+      ) {
+        setValidationError("Please fill all the fields");
+      } else {
+        setValidationError("");
+        setActiveStep(activeStep + 1);
+      }
+    }
+  };
+
+  const previousStep = (e) => {
+    if (activeStep !== steps.length) {
+      setActiveStep(activeStep - 1);
+    }
+  };
+
+  const hanldeConfirmationOfOrder = () => {
+    if (activeStep === 2) {
+      if (
+        customerInformation.paymentMethod === "" ||
+        customerInformation.cardNumber === "" ||
+        customerInformation.expiryDate === "" ||
+        customerInformation.cvv === ""
+      ) {
+        setValidationError("Please fill all the fields");
+      } else {
+        setValidationError("");
+        setActiveStep(activeStep + 1);
+      }
+    }
+  };
+
+  const getStepContent = (stepIndex) => {
+    switch (stepIndex) {
+      case 0:
+        return (
+          <Case0
+            validationError={validationError}
+            setCustomerInformation={setCustomerInformation}
+            customerInformation={customerInformation}
+          />
+        );
+
+      case 1:
+        return (
+          <Case1
+            validationError={validationError}
+            setCustomerInformation={setCustomerInformation}
+            customerInformation={customerInformation}
+          />
+        );
+
+      case 2:
+        return (
+          <Case2
+            validationError={validationError}
+            setCustomerInformation={setCustomerInformation}
+            customerInformation={customerInformation}
+          />
+        );
+      case 3:
+        return (
+          <Case3
+            CartItems={CartItems}
+            discountedPrice={discountedPrice}
+            handleVoucher={handleVoucher}
+            customerInformation={customerInformation}
+          />
+        );
+      default:
+        return "Unknown stepIndex";
+    }
+  };
+
+  return (
+    <Box className="checkout-main-box">
+      <Box>
+        <Box className="checkout-main-box-container">
+          <Stepper
+            className="checkout-stepper"
+            activeStep={activeStep}
+            alternativeLabel
+          >
+            {steps.map((label) => (
+              <Step key={label}>
+                <StepLabel>{label}</StepLabel>
+              </Step>
+            ))}
+          </Stepper>
+        </Box>
+        <br />
+        <Box className="checkout-content">
+          <Box className="checkout-content-inner">
+            <Box>{getStepContent(activeStep)}</Box>
+
+            <Box className="checkout-button-box">
+              <Button
+                disabled={activeStep === 0}
+                variant="contained"
+                color="primary"
+                onClick={(e) => previousStep(e)}
+              >
+                Previous
+              </Button>
+              {activeStep === steps.length - 1 ? (
+                <Button variant="contained" color="primary">
+                  <Link to="/finishorder" className="checkout-button-link">
+                    Place Order
+                  </Link>
+                </Button>
+              ) : activeStep === steps.length - 2 ? (
+                <Button
+                  variant="contained"
+                  color="primary"
+                  onClick={hanldeConfirmationOfOrder}
+                >
+                  Proceed To Confirm Your Order
+                </Button>
+              ) : (
+                <Button
+                  variant="contained"
+                  color="primary"
+                  onClick={(e) => nextStep(e)}
+                >
+                  Next
+                </Button>
+              )}
+            </Box>
+          </Box>
+        </Box>
+      </Box>
+    </Box>
+  );
+}
