+++ src/components/Shop.css
@@ -371,293 +371,11 @@
   margin-bottom: 1rem;
   color: #383838;
 }
+
 .product-review-createon {
   display: flex;
   text-align: center;
   align-items: center;
-}
-
-.customer-details {
-  display: flex;
-  flex-direction: column;
-  height: 300px;
-}
-
-.full-name {
-  padding: 3px;
-  border-radius: 5px;
-  width: 40%;
-  margin-top: 2rem;
-}
-
-.email {
-  padding: 3px;
-  border-radius: 5px;
-  width: 40%;
-  margin-top: 1rem;
-}
-
-.phone-number {
-  padding: 3px;
-  border-radius: 5px;
-  width: 40%;
-  margin-top: 1rem;
-}
-
-.shipping-details {
-  display: flex;
-  flex-direction: column;
-  height: 300px;
-}
-
-.shipping-address {
-  margin-top: 30px;
-  margin-bottom: 10px;
-}
-
-.street-address {
-  padding: 3px;
-  border-radius: 5px;
-  width: 40%;
-}
-
-.city-zip-code {
-  margin-top: 10px;
-  margin-bottom: 10px;
-}
-
-.city {
-  padding: 3px;
-  border-radius: 5px;
-  width: 40%;
-}
-
-.zip-code {
-  padding: 3px;
-  border-radius: 5px;
-  margin-left: 10px;
-  width: 40%;
-}
-
-.country-province {
-  margin-top: 10px;
-  margin-bottom: 10px;
-}
-
-.country {
-  padding: 3px;
-  border-radius: 5px;
-  width: 40%;
-}
-
-.province {
-  padding: 3px;
-  border-radius: 5px;
-  margin-left: 10px;
-  width: 40%;
-}
-
-.payment-details {
-  display: flex;
-  flex-direction: column;
-  height: 300px;
-}
-
-.payment-method {
-  margin-top: 10px;
-  color: black;
-  font-size: 1.3rem;
-}
-
-.card-number {
-  margin-top: 5px;
-  padding: 3px;
-  border-radius: 5px;
-  margin-bottom: 10px;
-  width: 50%;
-}
-
-.CVV {
-  padding: 3px;
-  border-radius: 5px;
-  margin-top: 10px;
-  margin-bottom: 10px;
-  width: 50%;
-}
-
-.expiry-date {
-  padding: 3px;
-  border-radius: 5px;
-  margin-bottom: 10px;
-  margin-top: 10px;
-  width: 50%;
-}
-
-.confirmation-of-order-main-box {
-  display: flex;
-  flex-direction: column;
-  height: auto;
-  width: auto;
-}
-
-.customer-details-box {
-  display: flex;
-  flex-direction: row;
-  width: 100vw;
-  justify-content: space-between;
-}
-
-.customer-details-box-inner {
-  width: auto;
-  display: flex;
-  flex-direction: column;
-  font-size: 18px;
-  margin-top: 10px;
-}
-
-.customer-details-items {
-  width: 75%;
-  display: flex;
-  flex-direction: column;
-  margin-top: 10px;
-}
-
-.customer-details-box-inner-address {
-  border-radius: 5px;
-  width: 20vw;
-  padding: 0.5rem;
-}
-
-.confirmation-of-order-button-box {
-  margin-top: 1rem;
-  font-size: large;
-  font-weight: 700;
-}
-
-.total-price-box {
-  font-size: 1.5rem;
-  font-weight: bold;
-  margin-bottom: 10px;
-  color: #0596f7;
-}
-
-.checkout-main-box {
-  position: relative;
-  background-size: cover;
-  height: 100vh;
-  background-color: rgb(243, 252, 255);
-}
-
-.checkout-main-box-container {
-  height: auto;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.checkout-stepper {
-  margin-top: 21px;
-  width: 70%;
-  color: white;
-  padding-top: 2px;
-}
-
-.checkout-content {
-  display: flex;
-  justify-content: center;
-}
-
-.checkout-content-inner {
-  border-radius: 14px;
-  margin-left: 14px;
-  padding-left: 14px;
-  margin-right: 14px;
-  padding-top: 12px;
-  box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.75);
-  background-color: rgba(255, 255, 255, 0.3);
-  width: 70%;
-}
-
-.checkout-button-box {
-  display: flex;
-  margin-top: 3px;
-  margin-bottom: 12px;
-  gap: 10px;
-}
-
-.checkout-button-link {
-  text-decoration: none;
-  color: white;
-}
-
-.category-list {
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-  margin-top: 5%;
-}
-
-.category-list-item {
-  width: auto;
-  height: auto;
-  margin-left: 0.5rem;
-  border-radius: 10px;
-  padding: 1rem;
-  background-color: #f5f5f5;
-}
-
-
-.product-list {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  width: auto;
-  height: 90vh;
-  gap: 30px;
-}
-
-.product-list-item {
-  width: 30%;
-  height: 75vh;
-  border-radius: 5px;
-  text-decoration: none;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  margin-right: 1rem;
-}
-
-.product-list-item:hover {
-  box-shadow: 0px 0px 5px #000000;
-  background-color: #f6f6f6;
-  border-radius: 5px;
-}
-
-.product-list-item-link {
-  text-decoration: none;
-  color: black;
-}
-
-.product-list-item-image {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  border-radius: 5px;
-  background-color: #f9f9f9;
-}
-
-.product-list-item-name {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 5px;
-}
-
-.product-list-item-name-text {
-  font-weight: bold;
 }
 
 .cart-modal {
